Autogenerate typings
Change-type: minor
