+++ index.d.ts
@@ -1,0 +1,5 @@
+// TODO: Drop this file once TS supports `export as namespace` in plain .ts files
+// See: https://github.com/microsoft/TypeScript/issues/10907
+export * from './es2015';
+// As generated by browserify
+export as namespace balenaSdk;
