+++ lib/models/index.ts
@@ -96,7 +96,11 @@ const modelsTemplate = {
 };
 
 export = (deps: InjectedDependenciesParam, opts: InjectedOptionsParam) => {
-	const models = {};
+	const models = {} as {
+		[key in keyof typeof modelsTemplate]: ReturnType<
+			ReturnType<typeof modelsTemplate[key]>
+		>;
+	};
 
 	(Object.keys(modelsTemplate) as Array<keyof typeof modelsTemplate>).forEach(
 		(moduleName) => {
