+++ lib/models/key.ts
@@ -15,8 +15,8 @@ limitations under the License.
 */
 
 import * as errors from 'balena-errors';
-import type * as BalenaSdk from '../..';
-import { InjectedDependenciesParam, InjectedOptionsParam } from '..';
+import type * as BalenaSdk from '..';
+import type { InjectedDependenciesParam, InjectedOptionsParam } from '..';
 import { mergePineOptions } from '../util';
 
 const getKeyModel = function (
