--- lib/types/os-types.ts
@@ -1,7 +1,0 @@
-/**
- * Extracted from the enum defined in `hostapp.ts`.
- */
-export enum OsTypes {
-	DEFAULT = 'default',
-	ESR = 'esr',
-}
