+++ tsconfig.json
@@ -1,7 +1,7 @@
 {
     "compilerOptions": {
         "module": "commonjs",
-        "declaration": false,
+        "declaration": true,
         "outDir": "build",
         "noImplicitAny": true,
         "removeComments": false,
